+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## v0.4.4 (not yet released)
 
+- **feature**: [you can now customize your story div container](https://github.com/phenixdigital/phx_live_storybook/issues/135)
 - **improvement**: [removed routes helpers](https://github.com/phenixdigital/phx_live_storybook/pull/137)
   (will help transition to Phoenix 1.7)
 - **improvement**: [pass connect params to story page](https://github.com/phenixdigital/phx_live_storybook/pull/130)